Backend: Tried to make assessment api work on the system. But hard luck. Frontend: Fixed errors of production build
Due to Internal Server Error after making any api request to Assessment Api at Backend due to invalid context injection at constructor. It is not possible to test the API. So, Here it is just a hands on to make it work but no luck.

For frontend all the production build errors are resolved before this commit.